Proyecto/requirements.txt
six
pillow
plantuml